remaining functions into classes
