+++ packages/hedera-wallet-snap/packages/snap/src/rpc/account/getAccountInfo.ts
@@ -22,7 +22,7 @@ import { AccountInfoQuery } from '@hashgraph/sdk';
 import { providerErrors } from '@metamask/rpc-errors';
 import { divider, heading, text } from '@metamask/snaps-ui';
 import _ from 'lodash';
-import { HederaServiceImpl } from '../../services/impl/hedera/HederaServiceImpl';
+import { HederaServiceImpl } from '../../services/impl/hedera/service/HederaServiceImpl';
 import { HederaClientFactory } from '../../snap/HederaClientFactory';
 import { SnapUtils } from '../../utils/SnapUtils';
 import { SnapState } from '../../snap/SnapState';
