+++ packages/hedera-wallet-snap/packages/snap/src/snap/SnapAccounts.ts
@@ -7,7 +7,7 @@ import { divider, heading, text } from '@metamask/snaps-ui';
 import {
   HederaServiceImpl,
   getHederaClient,
-} from '../services/impl/hedera/HederaServiceImpl';
+} from '../services/impl/hedera/service/HederaServiceImpl';
 import { KeyStore, SnapDialogParams, WalletSnapState } from '../types/state';
 import { StateUtils } from '../utils/StateUtils';
 import { CryptoUtils } from '../utils/CryptoUtils';
